Update index.html
This isnt complete but I thought it felt a little cleaner. Still need to add Borders and Dividers and clean it up. 

Added collapsible nav bar, A small logo (Not sure if that works yet or not) and an extra tab for Guides, but we can rename that later or add to it. I havent built out the guides.html so its an empty shell right now.